# apis/sentinel/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Sentinel API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/contexts/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a context by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contexts"
                ],
                "summary": "Get context by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "context ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/context.OutputDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a context by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contexts"
                ],
                "summary": "Delete a context by its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "context ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Polciy name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contexts"
                ],
                "summary": "Update a context by it's ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "context ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New name to be assigned to an existing context",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/context.Input"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/context.OutputDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
        "/v1/grants/resources/{resource_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Shows all Principal access to Target Resources managed through a context",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Grants"
                ],
                "summary": "Shows all Principal access to Target Resources managed through a context",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/grant.Output"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
        "/v1/grants/resources/{resource_id}/contexts/{context_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a grant for a context on a resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Grants"
                ],
                "summary": "Update a grant that permits a context on a resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "context ID",
                        "name": "context_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Details about the Grant to be added",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/grant.Input"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/grant.Output"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
...
        "/v1/principal/{principal_id}/authorization": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Shows all the permissions this principal has to provided target resources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Get authorization details about a principal.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Principal ID",
                        "name": "principal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Name of the permissions which allow access to the target",
                        "name": "permissions",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Name of the targtes to which a permission allows access",
                        "name": "targets",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Context through which authorization is determined",
                        "name": "context_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Limit your search results.",
                        "name": "depth",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include permissions that have deny permit fields set",
                        "name": "include_denied",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/authorization.Output"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "authorization.Details": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/resource.Attributes"
                },
                "id": {
                    "type": "string"
                },
                "relationships": {
                    "type": "object",
                    "$ref": "#/definitions/authorization.Relationships"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "resource"
                    ]
                }
            }
        },
        "authorization.Output": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/authorization.Details"
                    }
                }
            }
        },
        "authorization.Permissions": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/permission.Attributes"
                    }
                }
            }
        },
        "authorization.Relationships": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "object",
                    "$ref": "#/definitions/authorization.Permissions"
                }
            }
        },
        "context.Attributes": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "context.Details": {
            "type": "object",
            "required": [
                "attributes",
                "type"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/context.Attributes"
                },
                "id": {
                    "type": "string"
                },
                "relationships": {
                    "type": "object",
                    "$ref": "#/definitions/context.Relationships"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "context"
                    ]
                }
            }
        },
        "context.Input": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/context.InputDetails"
                }
            }
        },
        "context.InputDetails": {
            "type": "object",
            "required": [
                "attributes",
                "type"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/context.Attributes"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "context"
                    ]
                }
            }
        },
...
        "context.OutputDetails": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/context.Details"
                }
            }
        },
        "context.Relationship": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/context.Resource"
                    }
                }
            }
        },
        "context.Relationships": {
            "type": "object",
            "required": [
                "principals",
                "target_resources"
            ],
            "properties": {
                "principals": {
                    "type": "object",
                    "$ref": "#/definitions/context.Relationship"
                },
                "target_resources": {
                    "type": "object",
                    "$ref": "#/definitions/context.Relationship"
                }
            }
        },
        "context.Resource": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "resource"
                    ]
                }
            }
        },
        "grant.Attributes": {
            "type": "object",
            "required": [
                "with_grant"
            ],
            "properties": {
                "with_grant": {
                    "type": "boolean"
                }
            }
        },
        "grant.Data": {
            "type": "object",
            "required": [
                "id",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "grant"
                    ]
                }
            }
        },
        "grant.Details": {
            "type": "object",
            "required": [
                "attributes",
                "context",
                "principal",
                "type"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/grant.Attributes"
                },
                "context": {
                    "type": "object",
                    "$ref": "#/definitions/grant.Relationship"
                },
                "id": {
                    "type": "string"
                },
                "principal": {
                    "type": "object",
                    "$ref": "#/definitions/grant.Relationship"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "grant"
                    ]
                }
            }
        },
        "grant.Input": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/grant.InputDetails"
                }
            }
        },
        "grant.InputDetails": {
            "type": "object",
            "required": [
                "attributes",
                "type"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/grant.Attributes"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "grant"
                    ]
                }
            }
        },
        "grant.Output": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/grant.Details"
                    }
                }
            }
        },
        "grant.Relationship": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/grant.Data"
                }
            }
        },
...
        "permission.Attributes": {
            "type": "object",
            "required": [
                "name",
                "permitted"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "permitted": {
                    "type": "string"
                }
            }
        },
...
        "resource.Attributes": {
            "type": "object",
            "required": [
                "source_id"
            ],
            "properties": {
                "context_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                }
            }
        },
...
        "views.ErrView": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "source": {
                    "type": "object",
                    "$ref": "#/definitions/views.Source"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "views.Source": {
            "type": "object",
            "properties": {
                "parameter": {
                    "type": "string"
                },
                "pointer": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}